feat: add hooks features
Add hooks features for app install.

+++ libs/linglong/src/linglong/package_manager/package_manager.cpp
@@ -19,6 +19,7 @@
 #include "linglong/utils/command/env.h"
 #include "linglong/utils/configure.h"
 #include "linglong/utils/finally/finally.h"
+#include "linglong/utils/hooks.h"
 #include "linglong/utils/packageinfo_handler.h"
 #include "linglong/utils/serialize/json.h"
 #include "linglong/utils/transaction.h"
@@ -725,6 +726,12 @@ QVariantMap PackageManager::installFromLayer(const QDBusUnixFileDescriptor &fd,
                           << "after install" << packageRef.toString() << ":"
                           << ret.error().message();
           }
+
+          ret = executePostInstallHooks(*newRef);
+          if (!ret) {
+              taskRef.reportError(std::move(ret).error());
+              return;
+          }
       };
 
     auto refSpec =
@@ -753,6 +760,18 @@ QVariantMap PackageManager::installFromUAB(const QDBusUnixFileDescriptor &fd,
 {
     if (!fd.isValid()) {
         return toDBusReply(-1, "invalid file descriptor");
+    }
+
+    std::unique_ptr<utils::InstallHookManager> installHookManager =
+      std::make_unique<utils::InstallHookManager>();
+    auto ret = installHookManager->parseInstallHooks();
+    if (!ret) {
+        return toDBusReply(ret);
+    }
+
+    ret = installHookManager->executeInstallHooks(fd.fileDescriptor());
+    if (!ret) {
+        return toDBusReply(-1, "uab package signature verification failed.");
     }
 
     auto uabRet = package::UABFile::loadFromFile(fd.fileDescriptor());
@@ -1057,6 +1076,12 @@ QVariantMap PackageManager::installFromUAB(const QDBusUnixFileDescriptor &fd,
                                     "Failed to generate some cache.\n" + result.error().message());
                 return;
             }
+        }
+
+        auto ret = executePostInstallHooks(newAppRef);
+        if (!ret) {
+            taskRef.reportError(std::move(ret).error());
+            return;
         }
 
         transaction.commit();
@@ -1367,6 +1392,13 @@ void PackageManager::Install(PackageTask &taskContext,
         }
     }
 
+    auto ret = executePostInstallHooks(newRef);
+    if (!ret) {
+        taskContext.updateState(linglong::api::types::v1::State::Failed,
+                                "Failed to execute postInstall hooks.\n" + ret.error().message());
+        return;
+    }
+
     transaction.commit();
     taskContext.updateState(linglong::api::types::v1::State::Succeed,
                             "Install " + newRef.toString() + " success");
@@ -1635,6 +1667,11 @@ void PackageManager::Uninstall(PackageTask &taskContext,
         return;
     }
 
+    auto ret = executePostUninstallHooks(ref);
+    if (!ret) {
+        qWarning() << "failed to execute postUninstall hooks" << ret.error();
+    }
+
     transaction.commit();
 
     taskContext.updateState(linglong::api::types::v1::State::Succeed,
@@ -1809,6 +1846,12 @@ void PackageManager::Update(PackageTask &taskContext,
             return;
         }
 
+        ret = executePostInstallHooks(newRef);
+        if (!ret) {
+            qCritical() << "failed to execute post install hooks" << ret.error().message();
+            return;
+        }
+
         auto result = this->tryGenerateCache(newRef);
         if (!result) {
             taskContext.updateState(linglong::api::types::v1::State::Failed,
@@ -1908,8 +1951,21 @@ void PackageManager::pullDependency(PackageTask &taskContext,
                 return;
             }
 
+            auto ret = executePostInstallHooks(*runtime);
+            if (!ret) {
+                taskContext.updateState(linglong::api::types::v1::State::Failed,
+                                        LINGLONG_ERRV(ret).message());
+                return;
+            }
+
             transaction.addRollBack([this, runtimeRef = *runtime, module]() noexcept {
                 auto result = this->repo.remove(runtimeRef, module);
+                if (!result) {
+                    qCritical() << result.error();
+                    Q_ASSERT(false);
+                }
+
+                result = executePostUninstallHooks(runtimeRef);
                 if (!result) {
                     qCritical() << result.error();
                     Q_ASSERT(false);
@@ -1949,6 +2005,27 @@ void PackageManager::pullDependency(PackageTask &taskContext,
         if (isTaskDone(taskContext.subState())) {
             return;
         }
+
+        auto ret = executePostInstallHooks(*base);
+        if (!ret) {
+            taskContext.updateState(linglong::api::types::v1::State::Failed,
+                                    LINGLONG_ERRV(ret).message());
+            return;
+        }
+
+        transaction.addRollBack([this, baseRef = *base, module]() noexcept {
+            auto result = this->repo.remove(baseRef, module);
+            if (!result) {
+                qCritical() << result.error();
+                Q_ASSERT(false);
+            }
+
+            result = executePostUninstallHooks(baseRef);
+            if (!result) {
+                qCritical() << result.error();
+                Q_ASSERT(false);
+            }
+        });
     }
 
     transaction.commit();
@@ -2407,6 +2484,53 @@ auto PackageManager::GenerateCache(const QString &reference) noexcept -> QVarian
       .message = "",
     });
     return result;
+}
+
+utils::error::Result<void>
+PackageManager::executePostInstallHooks(const package::Reference &ref) noexcept
+{
+    LINGLONG_TRACE("execute post install hooks for: " + ref.toString());
+
+    std::unique_ptr<utils::InstallHookManager> installHookManager =
+      std::make_unique<utils::InstallHookManager>();
+    auto ret = installHookManager->parseInstallHooks();
+    if (!ret) {
+        return LINGLONG_ERR(ret);
+    }
+
+    auto layerDir = this->repo.getLayerDir(ref);
+    if (!layerDir) {
+        return LINGLONG_ERR(layerDir);
+    }
+
+    auto appPath = layerDir->absolutePath();
+
+    ret = installHookManager->executePostInstallHooks(ref.id.toStdString(), appPath.toStdString());
+    if (!ret) {
+        return LINGLONG_ERR(ret);
+    }
+
+    return LINGLONG_OK;
+}
+
+utils::error::Result<void>
+PackageManager::executePostUninstallHooks(const package::Reference &ref) noexcept
+{
+    LINGLONG_TRACE("execute post uninstall hooks for: " + ref.toString());
+
+    std::unique_ptr<utils::InstallHookManager> installHookManager =
+      std::make_unique<utils::InstallHookManager>();
+    auto ret = installHookManager->parseInstallHooks();
+    if (!ret) {
+        return LINGLONG_ERR(ret);
+    }
+
+    ret = installHookManager->executePostUninstallHooks(ref.id.toStdString());
+    if (!ret) {
+        return LINGLONG_ERR(ret);
+    }
+
+    return LINGLONG_OK;
 }
 
 } // namespace linglong::service
